Move tests for create_db from test_app_factory to test_orm
